feat: removed friend invitations
