+++ src/pages/Invitations/Invitations.scss
@@ -17,6 +17,25 @@
     width: 100%;
 }
 
+.tournament-invites__container {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-around;
+    margin: 0;
+    width: 40%;
+}
+
+.tournament-invites__invite {
+    align-items: center;
+    background: $dominant;
+    border-radius: 20px;
+    color: $secondary;
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+    padding: 0 1em;
+}
+
 @media (min-width: 768px) {
     .invitations__container {
         width: 80%;
