--- src/sections/FriendInvites/FriendInvites.jsx
@@ -1,13 +1,0 @@
-// import PropTypes from 'prop-types';
-
-export default function FriendInvites() {
-  return (
-    <section className="friend-invites__container">
-      <h1>Friend Requests</h1>
-    </section>
-  );
-}
-
-// FriendInvites.propTypes = {
-//   friendInvites: PropTypes.arrayOf(PropTypes.object).isRequired,
-// };
