--- src/sections/TournamentInvites/TournamentInvites.scss
@@ -1,20 +1,0 @@
-@import '../../App.scss';
-
-.tournament-invites__container {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-    margin: 0;
-    width: 40%;
-}
-
-.tournament-invites__invite {
-    align-items: center;
-    background: $dominant;
-    border-radius: 20px;
-    color: $secondary;
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-    padding: 0 1em;
-}
